Sphinx extension to generate mouse CLI usage
